JVAlertController: set property "contentSizeForViewInPopover" with size for popover.
On iOS < 7 it seems not sufficient to set the frame size of the view for the popover.
There is a special property available to indicate the size of the popover.
It is "contentSizeForViewInPopover". Its value will be ignored if the
popover controller calculates a value on its own but since no custom
UIPopoverController is used, there is no such calculation done.

Setting this property makes the popover use proper sizes.