JVAlertTransitionDelegate: use legacy animaction function in case the newer is unavailable.
On older iOS the function
animateWithDuration:delay:usingSpringWithDamping:initialSpringVelocity:options:animations:completion:
is unavailable. Hence there is a need to use a different one, even if the alternative method
does not provide the same level of customization of the animation start.

On older iOS the following function is used instead:
animateWithDuration:delay:options:animations:completion: