for popovers write "title" to the title attribute of a UINavigationController.
Instead of using an UILabel element, a wrapper of type UINavigationController is used.
This trick makes the title really look like a title and not like a button.

Elements within the content pane are sourrounded by an inset, created by the popover.
Only if the title is a title of this wrapping UINavigationController there is
no inset and no button style used when drawing the title.